fixes for sprite transform and Div

+++ include/gba/syscall.h
@@ -60,10 +60,10 @@ s32 Mod(s32 num, s32 denom);
 s32 ModArm(s32 denom, s32 num);
 #else
 // New GCC doesn't like us calling a function 'Mod', so we'll just inline them all.
-#define Div(num, denom)    ({((denom) != 0) ? ((num) / (denom)) : 0;})
-#define Mod(num, denom)    ({((denom) != 0) ? ((num) % (denom)) : 0;})
-#define DivArm(denom, num) ({((denom) != 0) ? ((num) / (denom)) : 0;})
-#define ModArm(denom, num) ({((denom) != 0) ? ((num) % (denom)) : 0;})
+#define Div(num, denom)    ({((denom) != 0) ? ((s32)(num) / (denom)) : 0;})
+#define Mod(num, denom)    ({((denom) != 0) ? ((s32)(num) % (denom)) : 0;})
+#define DivArm(denom, num) ({((denom) != 0) ? ((s32)(num) / (denom)) : 0;})
+#define ModArm(denom, num) ({((denom) != 0) ? ((s32)(num) % (denom)) : 0;})
 #endif
 
 void SoundBiasReset(void);
